Generate interesting Days and UniversalTimes
Previously the range of generated `Day` values was `size` days before and after 1864-5-8, i.e. about two months in the late nineteenth century. That's both too narrow (only covering part of the year) and doesn't intersect any of the dates which are likely to be of interest to most users of these types.

This makes the range `size` _years_, centered on 2000-1-1. With the typical size, that covers most of the era of modern computing and about a decade into the future.

Note: the Day instance is also used to generate LocalDates, UTCTimes, etc., so they also benefit.